requirements.txt
streamlit
pandas
dnspython
openpyxl
xlrd
